CS 1530/FunkyFantasyHosting/requirements.txt
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.3.23
Werkzeug==1.0.1
requests==2.27.1
pandas==1.3.5
espn_api==0.19.2
google-cloud-datastore
google-cloud-bigquery-storage
google-cloud-bigquery==2.3.1
google-cloud-storage
py-bcrypt
pyarrow